Blaming regex.
(auto-msg whatthecommit.com)

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   src/main.rs

+++ src/main.rs
@@ -95,7 +95,7 @@ async fn run() {
             let v = input_data[i];
             var finali = 0;
             for (var j = 0u; j < length_data; j++) {
-                if (input_data[j] == v && j < u32(i)) {
+                if (input_data[j] == v && j < i) {
                     finali += 1;
                 }
                 if (input_data[j] < v) {
